+++ tests/sample-data/column-chart.js
@@ -1,0 +1,31 @@
+'use strict';
+
+module.exports = [
+	{
+		id: 1,
+		date: '2020-04-15',
+		name: 'First element',
+		quantity: 10,
+		quantityColor: 'blue',
+		double: 20,
+		dobleAnnotation: 'A1'
+	},
+	{
+		id: 2,
+		date: '2020-04-16',
+		name: 'Second element',
+		quantity: 20,
+		quantityColor: '#b87333',
+		double: 40,
+		dobleAnnotation: 'A2'
+	},
+	{
+		id: 3,
+		date: '2020-04-17',
+		name: 'Third element',
+		quantity: 60,
+		quantityColor: 'color: #e5e4e2',
+		double: 120,
+		dobleAnnotation: 'A3'
+	}
+];