+++ CHANGELOG.md
@@ -7,12 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Added ColumnChart formatter
+- Implement `value` and `attributes` properties for label and values items.
+- Change functionality in helpers functions for make titles and values.
+
 ## [1.1.0] - 2020-05-05
+
 ### Added
+
 - Add BarChart formatter
 
 ## [1.0.0] - 2020-10-08
+
 ### Added
+
 - Project inited
 - API Dispatch
 - Fetcher